requirements.txt
requests
python-dotenv
pandas
psycopg2-binary
matplotlib
mplfinance
numpy
scipy
scikit-learn